refactor: custom components wallet spinup (#686)
* refactor: wallet spinup component


* update import type


* fix: resolved sonarlint issues


* fix: sonarcloud checks


---------
